Ensure key corresponds to the manga when not grouping
`v-for="(mg, key) in sortedMangas"` was misleading because
`sortedMangas` is an Array and `key` is thus an integer index. That
means that elements were reused depending on their position in the list,
not the actual manga. In particular, the flip-list transition was
defeated.

Use a key unique to the manga instead.