Experimental Komga Support
Adding experimental support for local reading using Komga. To use it you need to set up komga, and enter your url, username and password in the AMR options page.

Url for downloading the software: https://komga.org

This presented a few challenges to work through but I think it is stable enough for now that I want to push it out to get some feedback.

Note: Komga is a SPA so navigating to a chapter will not load amr even if you set it up in the options correctly, you will need to refresh on a chapter page to trigger AMR

* Also included is a change to how we remove styles, we continue to check for external stylesheets so any sites that load them dynamically will have them removed also