WIP marking series with update errors
This is a WIP update that adds a marking and list filter for series that have errors when updating the chapter list. It currently does not persist across browser restart or extension reloading for some reason. I will try and track it down when I move this to the V3 conversion